// matrix-meetings-bot/src/model/IUserContext.ts
/*
 * Copyright 2022 Nordeck IT + Consulting GmbH
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

// events and https-requests are containing information over the current user, who invokes them
// collect these data and use it as UserContext forfurther method-calls.
// can alter on be enriched by the unique-logging-id and any user relevant data,
export interface IUserContext {
  readonly locale: string;
  readonly timezone: string;
  readonly userId: string;
  //TODO: PB-1909 add contextType: 'http' | 'rpc'
  // to be used if different statements needed in services
  // for example: check in meeting service if user is in room, no need to apply for 'rpc'
  requestId?: string;
}